+++ GloryECS/GloryECSStatic/ComponentTypes.h
@@ -12,7 +12,7 @@ namespace GloryECS
 	{
 		ComponentType()
 			: m_Name("INVALID"), m_TypeHash(0), m_AllowMultiple(false), m_CustomFags(0) {}
-		ComponentType(const std::string& name, const size_t typeHash, const bool allowMultiple, const uint64_t customFlags)
+		ComponentType(const std::string& name, const uint32_t typeHash, const bool allowMultiple, const uint64_t customFlags)
 			: m_Name(name), m_TypeHash(typeHash), m_AllowMultiple(allowMultiple), m_CustomFags(customFlags) {}
 
 		const std::string m_Name;
@@ -29,7 +29,7 @@ namespace GloryECS
 		{
 			TypeView<T>* pTypeView = new TypeView<T>(nullptr);
 			std::type_index type = typeid(T);
-			size_t hash = Hashing::Hash(type.name());
+			uint32_t hash = Hashing::Hash(type.name());
 			m_pInstance->m_TypeHashes.push_back(hash);
 			std::string name = type.name();
 			AddTypeView(name, hash, pTypeView);
@@ -41,21 +41,21 @@ namespace GloryECS
 		static void SetInstance(ComponentTypes* pInstance);
 		static size_t GetComponentHash(const std::string& name);
 		static const size_t ComponentCount();
-		static const ComponentType* GetComponentType(const size_t hash);
+		static const ComponentType* GetComponentType(const uint32_t hash);
 		static const ComponentType* GetComponentTypeAt(const size_t index);
 
 	private:
-		static void AddTypeView(std::string& name, size_t hash, BaseTypeView* pTypeView);
-		static void AddComponentType(std::string& name, size_t hash, bool allowMultiple, uint64_t customFlags);
+		static void AddTypeView(std::string& name, uint32_t hash, BaseTypeView* pTypeView);
+		static void AddComponentType(std::string& name, uint32_t hash, bool allowMultiple, uint64_t customFlags);
 
 		template<typename T>
 		static TypeView<T>* CreateTypeView(EntityRegistry* pRegistry)
 		{
-			size_t hash = Hashing::Hash(typeid(T).name());
+			uint32_t hash = Hashing::Hash(typeid(T).name());
 			return CreateTypeView(pRegistry, hash);
 		}
 
-		static BaseTypeView* CreateTypeView(EntityRegistry* pRegistry, size_t hash);
+		static BaseTypeView* CreateTypeView(EntityRegistry* pRegistry, uint32_t hash);
 		static void ProcessName(std::string& name);
 
 	private:
